Support change_table_comment to change the table comment in migration
Enabled specify comment of database objects is introduced at #22911.
But for table comment, support only in the create_table clause,
so cannot add or modify comment for the existing table.
But add (modify) table comment is helpful to develop an app that
getting complex as time proceeds, or maintaining a historic app.